requirements.txt
fastapi
uvicorn[standard]
python-multipart
tensorflow
pydicom
pillow
numpy
pandas
scikit-learn
openpyxl
